requirements.txt
Django==5.0.2
mysqlclient==2.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
Pillow==10.3.0
djangorestframework-simplejwt==5.3.1